working button press for My Groups
